+++ src/app/core/auth/auth.service.ts
@@ -12,7 +12,7 @@ import { Address } from '../model/address';
 import { User } from '../model/user';
 import { environment } from '../../../environments/environment';
 import { StorageService } from './storage.service';
-import { CustomerService } from '../customer/customer.service';
+import { CustomerService } from '../customer/services/customer.service';
 import API_ENDPOINT from '../../shared/constants/api-endpoint';
 
 @Injectable({
