+++ src/app/core/category/services/category.api.service.ts
@@ -3,10 +3,10 @@ import { Injectable } from '@angular/core';
 
 import { map, Observable } from 'rxjs';
 
-import { environment } from '../../../environments/environment';
-import API_ENDPOINT from '../../shared/constants/api-endpoint';
-import { Category } from './interfaces/category';
-import { CategoryResponse } from './interfaces/category-response';
+import { environment } from '../../../../environments/environment';
+import API_ENDPOINT from '../../../shared/constants/api-endpoint';
+import { Category } from '../interfaces/category';
+import { CategoryResponse } from '../interfaces/category-response';
 
 @Injectable({
   providedIn: 'root',
