+++ src/app/core/customer/services/customer.service.ts
@@ -2,11 +2,10 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 
 import { Observable } from 'rxjs';
-
-import { CustomerAction } from './interfaces/customer-action';
-import { UserResponse } from '../auth/interfaces/user-response';
-import { environment } from '../../../environments/environment.dev';
-import API_ENDPOINT from '../../shared/constants/api-endpoint';
+import { environment } from '../../../../environments/environment';
+import API_ENDPOINT from '../../../shared/constants/api-endpoint';
+import { UserResponse } from '../../auth/interfaces/user-response';
+import { CustomerAction } from '../interfaces/customer-action';
 
 @Injectable({
   providedIn: 'root',
