+++ src/app/core/product/services/product-projections.helper.service.spec.ts
