+++ src/app/features/product/product-list/product-list.component.ts
@@ -6,12 +6,12 @@ import { Router, RouterLink } from '@angular/router';
 import { BOOKS_ID, COSMETICS_ID } from '../../../shared/constants/category';
 import { NavigateToSpecificRouteService } from '../../../shared/services/navigate-to-specific-route/navigate-to-specific-route.service';
 import { Category } from '../../../core/category/interfaces/category';
-import { CategoryApiService } from '../../../core/category/category.api.service';
-import { CategoryHelperService } from '../../../core/category/category.helper.service';
+import { CategoryApiService } from '../../../core/category/services/category.api.service';
+import { CategoryHelperService } from '../../../core/category/services/category.helper.service';
 import { ProductProjection } from '../../../core/product/interfaces/product-projection';
 import { ProductProjectionsResponse } from '../../../core/product/interfaces/product-projections-response';
-import { ProductProjectionsApiService } from '../../../core/product/product-projections.api.service';
-import { ProductProjectionsHelperService } from '../../../core/product/product-projections.helper.service';
+import { ProductProjectionsApiService } from '../../../core/product/services/product-projections.api.service';
+import { ProductProjectionsHelperService } from '../../../core/product/services/product-projections.helper.service';
 import { BriefCardComponent } from '../components/brief-card/brief-card.component';
 import { ProductButtonComponent } from '../components/product-button/product-button.component';
 import { SortByPriceComponent } from '../components/sort-by-price/sort-by-price.component';
